Version 1.1
Now if there is a package path classname such as java.util.Map
in ones code, then with cursor over Map, the Map.java file
is opened and if cursor is over for example 'util' then the
java/util directory is opened. All of the other means of opening
the java file under the cursor are still there.