Version 1.2
The entry point function, 'EditJava', now takes a command
argument which is executed if a matching file or directory is
found. This allows the user to parameterize the call with their
choice of opening a new window for the source file, e.g., the
command can be:
e, vsplit, new, split, vertical, sview, vnew , etc.

Also, the 'EditJava' method was refactored.